ClassCharacterError

+++ api/src/main/java/pl/kolendateam/dadcard/classCharacter/dto/ClassPcDTO.java
@@ -5,9 +5,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ClassPcDTO {
 
-    public short id;
-    public String className;
-    public int level;
-    public int[] classStudyId;
-
+  public int id;
+  public String className;
+  public int level;
+  public int[] classStudyId;
 }
