+++ api/src/main/java/pl/kolendateam/dadcard/classCharacter/repository/ClassRepository.java
@@ -1,13 +1,10 @@
 package pl.kolendateam.dadcard.classCharacter.repository;
 
 import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import pl.kolendateam.dadcard.classCharacter.entity.ClassCharacter;
 
-public interface ClassRepository extends JpaRepository<ClassCharacter, Short>{
-
-    Optional <ClassCharacter> findByName(String className);
-    
+public interface ClassRepository
+  extends JpaRepository<ClassCharacter, Integer> {
+  Optional<ClassCharacter> findByName(String className);
 }
